# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: AgentOps Desktop HTTP API
  description: |
    HTTP REST API for AgentOps Desktop — manage AI agents, goals, tasks, schedules, squads, cost budgets, adapters, and skills.

    **Base URL:** `http://localhost:3967`
    **Authentication:** Bearer token via `POST /auth/login`. Include as `Authorization: Bearer <token>` on all `/api/*` routes.
    **Default port:** 3967 (configurable via `PORT` env var)
  version: "0.2.0"
  contact:
    name: AgentOps Team
  license:
    name: UNLICENSED
servers:
  - url: "http://localhost:3967"
    description: Local HTTP API server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token from POST /auth/login
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          example: "task-1"
        title:
          type: string
          example: Implement auth
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, assigned, running, done, failed, blocked]
          example: pending
        goalId:
          type: string
          nullable: true
        assigneeAgentId:
          type: string
          nullable: true
        dependsOn:
          type: object
          nullable: true
        output:
          type: object
          nullable: true
        createdAt:
          type: integer
        updatedAt:
          type: integer
...
    Skill:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        content:
          type: string
        tags:
          type: object
          nullable: true
        createdAt:
          type: integer
        updatedAt:
          type: integer
paths:
...
  /api/tasks:
    get:
      summary: List tasks
      operationId: tasksList
      tags: [Tasks]
      security:
        - bearerAuth: []
      parameters:
        - { name: offset, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: status, in: query, schema: { type: string, enum: [pending, assigned, running, done, failed, blocked] } }
        - { name: goalId, in: query, schema: { type: string } }
        - { name: sortBy, in: query, schema: { type: string, enum: [createdAt, updatedAt, title, status] } }
        - { name: sortOrder, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Task" }
    post:
      summary: Create task
      operationId: tasksCreate
      tags: [Tasks]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, minLength: 1, maxLength: 500 }
                description: { type: string, maxLength: 5000 }
                goalId: { type: string }
                assigneeAgentId: { type: string }
                dependsOn: { type: object }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: "#/components/schemas/Task" }
  /api/tasks/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      summary: Get task
      operationId: tasksGet
      tags: [Tasks]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Task
        "404":
          description: Not found
    patch:
      summary: Update task
      operationId: tasksUpdate
      tags: [Tasks]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                status: { type: string, enum: [pending, assigned, running, done, failed, blocked] }
                goalId: { type: string }
                assigneeAgentId: { type: string }
                output: { type: object }
                dependsOn: { type: object }
      responses:
        "200":
          description: Updated
        "404":
          description: Not found
    delete:
      summary: Delete task
      operationId: tasksDelete
      tags: [Tasks]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found
  /api/tasks/{id}/upstream:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      summary: Get upstream task outputs
      operationId: tasksUpstream
      tags: [Tasks]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Upstream outputs
        "404":
          description: Task not found
  /api/tasks/{id}/handoffs:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      summary: List task handoffs
      operationId: tasksHandoffs
      tags: [Tasks]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Handoffs (outgoing + incoming)
        "404":
          description: Task not found

  # ── Logs ──
...
  /api/skills:
    get:
      summary: List skills
      operationId: skillsList
      tags: [Skills]
      security:
        - bearerAuth: []
      parameters:
        - { name: offset, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: tag, in: query, schema: { type: string } }
        - { name: tags, in: query, schema: { type: string, description: "Comma-separated tags" } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: Skill list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Skill" }
    post:
      summary: Create skill
      operationId: skillsCreate
      tags: [Skills]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, content]
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                description: { type: string, maxLength: 2000 }
                content: { type: string, minLength: 1 }
                tags: { type: object }
      responses:
        "201":
          description: Created
        "409":
          description: Skill name already exists
  /api/skills/tags:
    get:
      summary: List all tags
      operationId: skillsTags
      tags: [Skills]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tag list
  /api/skills/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      summary: Get skill
      operationId: skillsGet
      tags: [Skills]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Skill
        "404":
          description: Not found
    patch:
      summary: Update skill
      operationId: skillsUpdate
      tags: [Skills]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                content: { type: string }
                tags: { type: object }
      responses:
        "200":
          description: Updated
        "404":
          description: Not found
    delete:
      summary: Delete skill
      operationId: skillsDelete
      tags: [Skills]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found